+++ src/scripts/feed/PostEntry.js
@@ -1,25 +1,25 @@
 import { createNewPost } from "../data/provider.js";
 const applicationElement = document.querySelector(".giffygram");
 
-applicationElement.addEventListener("click", (clickEvent) => {
-  if (clickEvent.target.id === "newPost__submit") {
-    // Get what the user typed into the form fields
-    const titleHere = document.querySelector(
-      "input[name='postTitle']"
-    ).value;
-    const urlHere = document.querySelector(
-      "input[name='postURL']"
-    ).value;
-    const descriptionHere = document.querySelector(
-      "input[name='postDescription']"
-    ).value;
+// applicationElement.addEventListener("click", (clickEvent) => {
+//   if (clickEvent.target.id === "newPost__submit") {
+//     // Get what the user typed into the form fields
+//     const titleHere = document.querySelector(
+//       "input[name='postTitle']"
+//     ).value;
+//     const urlHere = document.querySelector(
+//       "input[name='postURL']"
+//     ).value;
+//     const descriptionHere = document.querySelector(
+//       "input[name='postDescription']"
+//     ).value;
 
-    const newPostFields = {
-      title: titleHere,
-      url: urlHere,
-      description: descriptionHere,
-    };
+//     const newPostFields = {
+//       title: titleHere,
+//       url: urlHere,
+//       description: descriptionHere,
+//     };
 
-    createNewPost(newPostFields);
-  }
-});
+//     createNewPost(newPostFields);
+//   }
+// });
